Holy shit.

+++ Box_Arduino/Box_Arduino.ino
@@ -84,7 +84,7 @@ typedef int (*Mapping)(long, int);
 Mapping mapping = &forward;
 
 
-int currentTime;
+unsigned long currentTime;
 unsigned long lastMillis;
 unsigned long internalTimeSmoother;
 
@@ -164,7 +164,7 @@ void read() {
         for (int i = 0; i < sub.length(); i++) {
           c[i] = sub.charAt(i);
         }
-        currentTime = atoi(c);
+        currentTime = atol(c);
 
         Serial.print("Current time: ");
         Serial.println(currentTime);
